did the boughtBy functionality
There was a problem with the "buy" functionality which I fixed. I needed to use the .valueOf() on every element of the mongoose reference collection of ID's because otherwise they display as ObjectId('645avf8982fyh9876') and it can not be compared to normal '645avf8982fyh9876', which means that every time my hasBought variable returned false!